requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
idna==3.4
lxml==4.9.2
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.1.1
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
requests-pkcs12==1.15
sqlparse==0.4.3
urllib3==1.26.15
